Add routine to pause DBImpl::Recover while new overlapped areas are cleared via compactions (Riak 1.3 to 1.4 conversion).  Also cleaned some int versus size_t compiler warnings.

+++ db/dbformat.h
@@ -31,7 +31,7 @@ static const size_t kL0_CompactionTrigger = 4;
 static const size_t kL0_SlowdownWritesTrigger = 8;
 
 // Maximum number of level-0 files.  We stop writes at this point.
-static const int kL0_StopWritesTrigger = 12;
+static const size_t kL0_StopWritesTrigger = 12;
 
 // Maximum level to which a new compacted memtable is pushed if it
 // does not create overlap.  We try to push to level 2 to avoid the
